add userId in model, update PUT and DELETE routes

+++ back/controllers/sauces.js
@@ -29,27 +29,44 @@ exports.createSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    const sauceObject = req.file ?
-    {
-        ...JSON.parse(req.body.sauce),
-        imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
-    } : { ...req.body };
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Sauce modified."}))
-    .catch(error => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId !== req.auth.userId) {
+                res.status(400).json({ error : "Unauthorized request." })
+            }
+            else if (sauce.userId == req.auth.userId) {
+                const filename = sauce.imageUrl.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+                    const sauceObject = req.file ?
+                        {
+                            ...JSON.parse(req.body.sauce),
+                            imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+                        } : { ...req.body };
+                    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce modified." }))
+                        .catch(error => res.status(400).json({ error }));
+                })
+            }
+        })
+        .catch(error => res.status(400).json({ error }));
 };
 
 exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({ _id: req.params.id })
-    .then(sauce => {
-        const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-            Sauce.deleteOne({ _id: req.params.id })
-            .then(() => res.status(200).json({ message: "Sauce deleted."}))
-            .catch(error => res.status(400).json({ error }));
-        });
-    })
-    .catch(error => res.status(500).json({ error }));
+        .then(sauce => {
+            if (sauce.userId !== req.auth.userId) {
+                res.status(400).json({ error: "Unauthorized request." })
+            }
+            else if (sauce.userId == req.auth.userId) {
+                const filename = sauce.imageUrl.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => res.status(200).json({ message: "Sauce deleted." }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch(error => res.status(500).json({ error }));
 };
 
 exports.likeSauce = (req, res, next) => {
